pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "bit>=0.8.0",
    "cryptography>=46.0.5",
    "discord-py>=2.7.1",
    "duckduckgo-search>=8.1.1",
    "feedparser>=6.0.12",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "qrcode[pil]>=8.2",
    "requests>=2.32.0",
    "solders>=0.27.1",
    "ta>=0.11.0",
    "telegram>=0.0.1",
    "web3>=7.14.1",
    "youtube-transcript-api>=1.2.4",
]
